+++ src/Service/CognitoIdentityProvider/src/Input/AdminCreateUserRequest.php
@@ -205,7 +205,7 @@ final class AdminCreateUserRequest extends Input
 
         // Prepare Body
         $bodyPayload = $this->requestBody();
-        $body = empty($bodyPayload) ? '{}' : json_encode($bodyPayload);
+        $body = empty($bodyPayload) ? '{}' : \json_encode($bodyPayload, 4194304);
 
         // Return the Request
         return new Request('POST', $uriString, $query, $headers, StreamFactory::create($body));