+++ src/Service/RdsDataService/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: With the Data API, you can now use UUID and JSON data types as input to your database.
+
 ## 0.1.2
 
 ### Added