+++ src/Service/RdsDataService/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Fixed
+
+- Assert the provided Input can be json-encoded.
+
 ### Added
 
 - AWS api-change: With the Data API, you can now use UUID and JSON data types as input to your database.